client: Fixed Visual Studio compiler warnings. In particular: - Changed off_t type to int64_t in interface. - Using size_t as index type where necessary. - Added static_cast where signed and unsigned integers are used together. - Replaced boost::*int_t* types with those from stdint.h
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3585 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ cpp/include/libxtreemfs/interrupt.h
@@ -8,6 +8,8 @@
 #ifndef CPP_INCLUDE_LIBXTREEMFS_INTERRUPT_H_
 #define CPP_INCLUDE_LIBXTREEMFS_INTERRUPT_H_
 
+#include <stdint.h>
+
 #include <boost/function.hpp>
 
 namespace xtreemfs {
@@ -26,7 +28,7 @@ class Interruptibilizer {
    * @remarks this function contains a boost::thread interruption point and
    *          thus might throw boost::thread_interrupted.
    */
-  static void SleepInterruptible(int rel_time_ms, InterruptedCallback cb);
+  static void SleepInterruptible(int64_t rel_time_ms, InterruptedCallback cb);
 };
 
 }  // namespace xtreemfs
